requirements.txt
# 核心依赖
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
duckduckgo-search>=8.0.0
ddgs>=9.0.0